added bias, not compiling yet

+++ include/pam_mujoco/run_mujoco.hpp
@@ -112,16 +112,16 @@ namespace pam_mujoco
 
   template<int NB_DOFS>
   void add_pressure_controller(std::string segment_id,
-			       int scale_min_pressure, int scale_max_pressure,
-			       int scale_min_activation, int scale_max_activation,
+			       double scale_min_pressure, double scale_max_pressure,
+			       double scale_min_activation, double scale_max_activation,
 			       std::string muscle_json_config_path_ago,
 			       std::string muscle_json_config_path_antago,
 			       std::array<double,NB_DOFS*2> a_init,
 			       std::array<double,NB_DOFS*2> l_MTC_change_init);
 
   void add_4dofs_pressure_controller(std::string segment_id,
-				     int scale_min_pressure, int scale_max_pressure,
-				     int scale_min_activation, int scale_max_activation,
+				     double scale_min_pressure, double scale_max_pressure,
+				     double scale_min_activation, double scale_max_activation,
 				     std::string muscle_json_config_path_ago,
 				     std::string muscle_json_config_path_antago,
 				     std::array<double,8> a_init,
